requirements.txt
beautifulsoup4
chromadb
fastapi
langchain
langchain-openai
langserve
python-dotenv
sse-starlette
streamlit
uvicorn